PlayerJS refactor relates to xibosignageltd/xibo-private#505
