+++ ui/src/helpers/date-format-helper.js
@@ -1,6 +1,8 @@
 const DateFormatHelper = function(options) {
   this.timezone = null;
   this.systemFormat = 'Y-m-d H:i:s';
+  this.macroRegex = /^%(\+|\-)[0-9]([0-9])?(d|h|m|s)%$/gi;
+
 
   this.convertPhpToMomentFormat = function(format) {
     if (String(format).length === 0) {
@@ -57,6 +59,33 @@ const DateFormatHelper = function(options) {
     });
 
     return convertedFormat;
+  };
+
+  this.composeUTCDateFromMacro = function(macroStr) {
+    const utcFormat = 'YYYY-MM-DDTHH:mm:ssZ';
+    const dateNow = moment().utc();
+    // Check if input has the correct format
+    const dateStr = String(macroStr);
+
+    if (dateStr.length === 0 ||
+        dateStr.match(this.macroRegex) === null
+    ) {
+      return dateNow.format(utcFormat);
+    }
+
+    // Trim the macro date string
+    const dateOffsetStr = dateStr.replaceAll('%', '');
+    const params = (op) => dateOffsetStr.replace(op, '')
+      .split(/(\d+)/).filter(Boolean);
+    const addRegex = /^\+/g;
+    const subtractRegex = /^\-/g;
+
+    // Check if it's add or subtract offset and return composed date
+    if (dateOffsetStr.match(addRegex) !== null) {
+      return dateNow.add(...params(addRegex)).format(utcFormat);
+    } else if (dateOffsetStr.match(subtractRegex) !== null) {
+      return dateNow.subtract(...params(subtractRegex)).format(utcFormat);
+    }
   };
 
   return this;
